+++ Makefile
@@ -1,14 +1,14 @@
 regular:
-	gcc -o tinker.out consume_memory.c tmalloc.c
+	gcc -o tinker.out consume_memory.c node.c tmalloc.c
 
 tinker:
-	gcc -D TINKER -o tinker.out consume_memory.c tmalloc.c
+	gcc -D TINKER -o tinker.out consume_memory.c node.c tmalloc.c
 
 debug_regular:
-	gcc -g -o debug.out consume_memory.c tmalloc.c
+	gcc -g -o debug.out consume_memory.c node.c tmalloc.c
 
 debug_tinker:
-	gcc -g -D TINKER -o debug.out consume_memory.c tmalloc.c
+	gcc -g -D TINKER -o debug.out consume_memory.c node.c tmalloc.c
 
 run:
 	ulimit -S -v 10240; ./tinker.out 10
